Use C interface of @salomonl
